Refactor insertion into BSP tree.
Before this commit, inserting into BSP tree could easily overflow
the stack because we allocate very large stack frames and, on
convex geometries (e.g. a sphere), the BSP tree degenerates into
a "BSP list", thus requiring one large stack frame per triangle.
This can be reproduced by exporting a 2d shaded view of sphere.

After this commit, the stack frames only contan a pointer to
a supplementary data structure, and moreover it only allocates
its fields on demand, conserving heap memory as well.

As a side effect, an arbitrary classifier limit of 50 vertices
is removed.

+++ src/polygon.h
@@ -236,9 +236,9 @@ public:
     void Insert(STriangle *str, SMesh *instead);
     static SBsp3 *InsertOrCreate(SBsp3 *where, STriangle *str, SMesh *instead);
 
-    void InsertConvexHow(BspClass how, STriMeta meta, Vector *vertex, int n,
+    void InsertConvexHow(BspClass how, STriMeta meta, Vector *vertex, size_t n,
                                 SMesh *instead);
-    SBsp3 *InsertConvex(STriMeta meta, Vector *vertex, int n, SMesh *instead);
+    SBsp3 *InsertConvex(STriMeta meta, Vector *vertex, size_t n, SMesh *instead);
 
     void InsertInPlane(bool pos2, STriangle *tr, SMesh *m);
